requirements.txt
absl-py==0.9.0
asgiref==3.2.10
astroid==2.3.1
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
class-registry==2.1.2
colorama==0.4.1
conda==4.8.3
conda-package-handling==1.7.0
cryptography==2.9.2
Django==3.0.8
django-cors-headers==3.4.0
django-filter==1.1.0
djangorestframework==3.11.0
djangorestframework-filters==0.11.1
djangorestframework-simplejwt==4.4.0
filters==1.3.2
gast==0.3.3
google-auth==1.19.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
idna==2.9
importlib-metadata==1.7.0
isort==4.3.21
jsonfield==3.1.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.4.2
Markdown==3.2.2
mccabe==0.6.1
menuinst==1.4.16
numpy==1.19.1
oauthlib==3.1.0
opt-einsum==3.3.0
Pillow==7.2.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.20
PyJWT==1.7.1
pylint==2.4.2
pyOpenSSL==19.1.0
PySocks==1.7.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
pywin32==227
PyYAML==5.3.1
regex==2020.7.14
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel-yaml==0.15.87
scipy==1.4.1
six==1.12.0
sqlparse==0.3.1
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
tqdm==4.46.0
typed-ast==1.4.0
urllib3==1.25.8
Werkzeug==1.0.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
zipp==3.1.0
